stripe_test/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
cryptography==44.0.1
django-filter==25.1
pillow==11.1.0
requests==2.32.3
django-extensions==3.2.3
python-decouple==3.8
stripe==11.6.0
django-cors-headers==4.7.0
gunicorn==23.0.0
